+++ .github/workflows/refactor.yml
@@ -1,13 +1,11 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Python package
+name: Refactor
 
 on:
   push:
-    branches: [main, development]
-  pull_request:
-    branches: [main, development]
+    branches: [project_refactor]
 
 jobs:
   build:
@@ -18,7 +16,7 @@ jobs:
         # 3.8 Shared hosting
         # 3.9 Ubi minimal
         # 3.12 Last release
-        python-version: [3.8, 3.9, 3.12]
+        python-version: [3.8]
 
     steps:
       - uses: actions/checkout@v3
