+++ now_lms/config/parse_config_file.py
@@ -33,7 +33,6 @@ from configobj import ConfigObj
 # ---------------------------------------------------------------------------------------
 from now_lms.logs import log
 
-
 # ---------------------------------------------------------------------------------------
 # Configuration can be loaded from a init like life.
 # ---------------------------------------------------------------------------------------
