+++ now_lms/templates/admin/users.html
@@ -24,7 +24,7 @@
 
         <p>
           <span class="badge bg-success">
-            <a href="{{ url_for("crear_usuario") }}" class="link-light">Nuevo Usuario</a>
+            <a href="{{ url_for("user.crear_usuario") }}" class="link-light">Nuevo Usuario</a>
           </span>
         </p>
 
@@ -51,7 +51,7 @@
             <tr>
 
               <td>
-                <a href="{{ url_for("usuario", id_usuario=item.usuario) }}" class="link-dark">
+                <a href="{{ url_for("admin_profile.usuario", id_usuario=item.usuario) }}" class="link-dark">
                   {{ item.usuario }}
                 </a>
               </td>
