+++ now_lms/templates/auth/login.html
@@ -34,7 +34,7 @@
   <main class="form-signin">
     <form method="post">
       {{ form.csrf_token }}
-      <a href="{{ url_for('home') }}">
+      <a href="{{ url_for('home.pagina_de_inicio') }}">
         <img class="mb-4" src="{{ url_for('static', filename='icons/logo/logo.svg') }}" alt="NOW - LMS" width="200">
       </a>
       <h1 class="h3 mb-3 fw-normal">Inicio de Sesión</h1>
@@ -52,7 +52,7 @@
       <label for="contraseña" class="visually-hidden">Contraseña</label>
       {{ form.acceso(class="form-control", id="acceso", placeholder="Contraseña") }}
       <button class="w-100 btn btn-lg btn-primary btn-block" type="submit">Iniciar Sesión</button>
-      <a href="{{ url_for('home') }}"><i class="bi bi-arrow-left" aria-hidden="true"></i> Regresar</a>
+      <a href="{{ url_for('home.pagina_de_inicio') }}"><i class="bi bi-arrow-left" aria-hidden="true"></i> Regresar</a>
       <p class="mt-5 mb-3 text-muted text-center">BMO Soluciones &copy; 2023</p>
     </form>
   </main>
