+++ now_lms/templates/error_pages/402.html
@@ -5,7 +5,7 @@
 
 <head>
   {{ macros.headertags() }}
-  <title>Permisos insuficientes.</title>
+  <title>Pago requerido.</title>
 
 </head>
 
@@ -20,9 +20,9 @@
     </header>
 
     <main class="px-3">
-      <h1><i class="bi bi-incognito" aria-hidden="true"></i></h1>
-      <h1>Perfil Privado</h1>
-      <p class="lead">El perfil del usuario solicitado es privado.</p>
+      <h1><i class="bi bi-credit-cardit" aria-hidden="true"></i></h1>
+      <h1>402</h1>
+      <p class="lead">Para acceder al recurso solicitado primero debe registrar un pago.</p>
       <p class="lead">
         <a class="btn btn-warning fw-bold" href="{{ url_for("home.pagina_de_inicio") }}"><i class="bi bi-house"
             aria-hidden="true"></i>
