+++ now_lms/templates/error_pages/405.html
@@ -20,9 +20,9 @@
     </header>
 
     <main class="px-3">
-      <h1><i class="bi bi-incognito" aria-hidden="true"></i></h1>
-      <h1>Perfil Privado</h1>
-      <p class="lead">El perfil del usuario solicitado es privado.</p>
+      <h1><i class="bi bi-slash-circle" aria-hidden="true"></i></h1>
+      <h1>403</h1>
+      <p class="lead">Su usuario no se encuentra autorizado a acceder al recurso solicitado.</p>
       <p class="lead">
         <a class="btn btn-warning fw-bold" href="{{ url_for("home.pagina_de_inicio") }}"><i class="bi bi-house"
             aria-hidden="true"></i>
