+++ now_lms/templates/error_pages/500.html
@@ -1,0 +1,43 @@
+{% import "macros.html" as macros %}
+
+<!DOCTYPE html>
+<html lang="es" class="h-100">
+
+<head>
+  {{ macros.headertags() }}
+  <title>Error al procesar su solicitud.</title>
+
+</head>
+
+<body class="d-flex h-100 text-center">
+
+  <div class="cover-container d-flex w-100 h-100 p-3 mx-auto flex-column">
+    <header class="mb-auto">
+      <a class="navbar-brand col-md-3 col-lg-2 me-0 px-3" href="{{ url_for("home.pagina_de_inicio") }}">
+        <img height="80" src="{{ url_for('static', filename='icons/logo/logo_large.png') }}" alt="NOW LMS"
+          class="d-inline-block align-top">
+      </a>
+    </header>
+
+    <main class="px-3">
+      <h1><i class="bi bi-bug" aria-hidden="true"></i></h1>
+      <h1>500</h1>
+      <p class="lead">A ocurrido un error al intentar atender su solicitud.</p>
+      <p class="lead">
+        <a href="{{ url_for("home.pagina_de_inicio") }}" class="btn btn-warning fw-bold"><i class="bi bi-house"
+            aria-hidden="true"></i>
+          Regresar al inicio</a>
+      </p>
+    </main>
+
+    <footer class="mt-auto">
+      <p>BMO Soluciones</p>
+      <p>&copy; 2021</p>
+    </footer>
+  </div>
+
+
+
+</body>
+
+</html>
