+++ now_lms/templates/inicio/cursos.html
@@ -82,7 +82,7 @@
                   <div class="d-flex justify-content-between align-items-center">
 
                     <div class="btn-group">
-                      <a href="{{ url_for("curso", course_code=curso.codigo )}}"
+                      <a href="{{ url_for("course.curso", course_code=curso.codigo )}}"
                         class="btn btn-sm btn-outline-secondary" role="button" aria-disabled="true">
                         Ver Curso
                       </a>
@@ -103,7 +103,7 @@
           </div>
 
           <div class="container" id="user_paginatios">
-            {{ macros.rendizar_paginacion(consulta=cursos, vista="lista_cursos") }}
+            {{ macros.rendizar_paginacion(consulta=cursos, vista="course.lista_cursos") }}
           </div>
 
         </div>
