+++ now_lms/templates/inicio/mooc.html
@@ -75,8 +75,8 @@
                 <div class="d-flex justify-content-between align-items-center">
 
                   <div class="btn-group">
-                    <a href="{{ url_for("curso", course_code=curso.codigo )}}" class="btn btn-sm btn-outline-secondary"
-                      role="button" aria-disabled="true">
+                    <a href="{{ url_for("course.curso", course_code=curso.codigo )}}"
+                      class="btn btn-sm btn-outline-secondary" role="button" aria-disabled="true">
                       Ver Curso
                     </a>
                   </div>
@@ -99,12 +99,12 @@
     <div class="text-center">
       <p>{{ cursos.total }} cursos disponibles.</p>
       {% if cursos.has_prev %}
-      <a class="link-dark" href="{{ url_for("home", page=cursos.prev_num) }}"><i class="bi bi-arrow-left-short"
-          aria-hidden="true"></i>Anterior</a>
+      <a class="link-dark" href="{{ url_for("home.pagina_de_inicio", page=cursos.prev_num) }}"><i
+          class="bi bi-arrow-left-short" aria-hidden="true"></i>Anterior</a>
       {% endif %}
       Página {{ cursos.page }} de {{ cursos.pages }}
       {% if cursos.has_next %}
-      <a class="link-dark" href="{{ url_for("home", page=cursos.next_num) }}">Siguiente <i
+      <a class="link-dark" href="{{ url_for("home.pagina_de_inicio", page=cursos.next_num) }}">Siguiente <i
           class="bi bi-arrow-right-short" aria-hidden="true"></i></a>
       {% endif %}
     </div>
