+++ now_lms/templates/learning/certificados/certificado.html
@@ -1,0 +1,26 @@
+{% import "macros.html" as macros %}
+
+<!DOCTYPE html>
+<html lang="es" class="h-100">
+
+<head>
+
+  {{ macros.headertags() }}
+  <title>{{ title }}</title>
+
+  {{ macros.local_style() }}
+</head>
+
+<body>
+
+  {{ macros.navbar() }}
+
+  <main>
+
+
+  </main>
+
+
+</body>
+
+</html>
