+++ now_lms/templates/learning/certificados/nuevo_certificado.html
@@ -1,0 +1,49 @@
+{% import "macros.html" as macros %}
+
+<!DOCTYPE html>
+<html lang="es" class="h-100">
+
+<head>
+
+  {{ macros.headertags() }}
+  <title>{{ title }}</title>
+
+  {{ macros.local_style() }}
+</head>
+
+<body>
+
+  {{ macros.navbar() }}
+
+  <main>
+
+    {{ macros.notify() }}
+
+    <div class="container px-4 py-5" id="icon-grid">
+      <h4 class="pb-2 border-bottom">Crear nuevo Certificado.</h4>
+      <form method="post" enctype=multipart/form-data>
+        {{ form.csrf_token }}
+        <div class="row g-3">
+          <div class="col-sm-6">
+            <label for="titulo" class="form-label">Certificado</label>
+            {{ form.titulo(class="form-control", id="titulo", placeholder="Certificado") }}
+          </div>
+
+          <div class="col-sm-6">
+            <label for="descripcion" class="form-label">Descripcion</label>
+            {{ form.descripcion(class="form-control", id="descripcion", placeholder="Descripcion") }}
+          </div>
+        </div>
+
+        <br>
+
+        <button class="w-10 btn btn-lg btn-primary btn-block" type="submit">Crear Certificado</button>
+      </form>
+    </div>
+
+  </main>
+
+
+</body>
+
+</html>
