+++ now_lms/templates/perfiles/admin.html
@@ -45,15 +45,15 @@
               Usuarios
             </h4>
             <p>
-              <a href="{{ url_for("usuarios") }}" class="link-dark">Usuarios</a>
+              <a href="{{ url_for("admin_profile.usuarios") }}" class="link-dark">Usuarios</a>
               -
-              <a href="{{ url_for("usuarios_inactivos") }}" class="link-dark">Usuarios Inactivos
+              <a href="{{ url_for("admin_profile.usuarios_inactivos") }}" class="link-dark">Usuarios Inactivos
                 {% if inactivos > 0 %}
                 ({{ inactivos | int }})
                 {% endif %}
               </a>
               -
-              <a href="{{ url_for("crear_usuario") }}" class="link-dark">Nuevo Usuario</a>
+              <a href="{{ url_for("user.crear_usuario") }}" class="link-dark">Nuevo Usuario</a>
             </p>
           </div>
         </div>
@@ -65,8 +65,8 @@
               Grupos
             </h4>
             <p>
-              <a href="{{ url_for("lista_grupos") }}" class="link-dark">Grupos</a> - <a
-                href="{{ url_for("nuevo_grupo")  }}" class="link-dark">Nuevo Grupo</a>
+              <a href="{{ url_for("instructor_profile.lista_grupos") }}" class="link-dark">Grupos</a> - <a
+                href="{{ url_for("group.nuevo_grupo")  }}" class="link-dark">Nuevo Grupo</a>
             </p>
             <p>
 
@@ -81,9 +81,9 @@
               Configuración
             </h4>
             <p>
-              <a href="{{ url_for("configuracion") }}" class="link-dark">Configuración del Sitio</a>
+              <a href="{{ url_for("setting.configuracion") }}" class="link-dark">Configuración del Sitio</a>
               -
-              <a href="{{ url_for("personalizacion") }}" class="link-dark">Tema del Sitio</a>
+              <a href="{{ url_for("setting.personalizacion") }}" class="link-dark">Tema del Sitio</a>
             </p>
           </div>
         </div>
@@ -96,7 +96,7 @@
               Correo electronico
             </h4>
             <p>
-              <a href="{{ url_for("mail") }}" class="link-dark">Configuración</a>
+              <a href="{{ url_for("setting.mail") }}" class="link-dark">Configuración</a>
             </p>
           </div>
         </div>
@@ -108,7 +108,7 @@
               Pagos
             </h4>
             <p>
-              <a href="{{ url_for("mail") }}" class="link-dark">Configuración</a>
+              <a href="{{ url_for("setting.mail") }}" class="link-dark">Configuración</a>
             </p>
           </div>
         </div>
