+++ now_lms/templates/perfiles/moderador.html
@@ -43,7 +43,7 @@
           <div>
             <h4 class="fw-bold mb-0">Mensajes</h4>
             <p>
-              <a href=" {{ url_for("cursos") }} " class="link-dark">Mensajes</a>
+              <a href=" {{ url_for("instructor_profile.cursos") }} " class="link-dark">Mensajes</a>
             </p>
           </div>
         </div>
