+++ now_lms/vistas/README.md
@@ -1,0 +1,6 @@
+# Now Learning Management System
+
+La logica del sistema de aprendizage esta divida en modulos utilizando
+[blueprints](https://flask.palletsprojects.com/en/2.3.x/blueprints/).
+Esto permite dividir la logica de la aplicación en unidades mas pequeñas
+y faciles de mantener.
